Change management of buttons on ModInstallMessageForm
To align OK/Cancel/Next buttons and other action buttons.
Next button should be the rightest, but OK button is placed left of
action buttons on Windows.
So we can't OK button as Next. Added Next button as an action button and
other action buttons are pre-defined and `setVisible` depending on each
dialog, to place Next rightest.